+++ lib/runner-helpers/multi-process/run-queue.ts
@@ -1,0 +1,26 @@
+'use strict';
+
+//dts
+import {IGlobalSumanObj} from "suman-types/dts/global";
+
+//polyfills
+const process = require('suman-browser-polyfills/modules/process');
+const global = require('suman-browser-polyfills/modules/global');
+
+//npm
+import async = require('async');
+
+//project
+const _suman: IGlobalSumanObj = global.__suman = (global.__suman || {});
+
+//////////////////////////////////////////////////////////////////////
+
+export const makeRunQueue = function () {
+
+  const {sumanConfig, maxProcs} = _suman;
+
+  return async.queue(function (task, cb) {
+
+  }, maxProcs);
+
+};
