+++ test/.suman/gantt-4.html
@@ -220,7 +220,7 @@
       "status": "KILLED"
     }];
 
-  let tasksRaw = JSON.parse('[{"startDate":1508379202647,"endDate":1508379203242,"transformStartDate":1508379202502,"transformEndDate":1508379202636,"taskName":"/Users/alexamil/WebstormProjects/oresoftware/sumanjs/suman/test/src/exp/tap-output/@src/tap-producer2.ts","status":"FAILED"},{"startDate":1508379202653,"endDate":1508379203243,"transformStartDate":1508379202506,"transformEndDate":1508379202637,"taskName":"/Users/alexamil/WebstormProjects/oresoftware/sumanjs/suman/test/src/exp/tap-output/@src/tap-producer1.ts","status":"FAILED"}]');
+  let tasksRaw = JSON.parse('[{"startDate":1508379280594,"endDate":1508379281139,"transformStartDate":1508379280455,"transformEndDate":1508379280584,"taskName":"/Users/alexamil/WebstormProjects/oresoftware/sumanjs/suman/test/src/exp/tap-output/@src/tap-producer2.ts","status":"FAILED"},{"startDate":1508379280598,"endDate":1508379281160,"transformStartDate":1508379280459,"transformEndDate":1508379280585,"taskName":"/Users/alexamil/WebstormProjects/oresoftware/sumanjs/suman/test/src/exp/tap-output/@src/tap-producer1.ts","status":"FAILED"}]');
 
   let tasks = [];
